Fix default eos_token and pad_token and 16-bit model loading.
### Tokens
This resolves default values for eos_token and pad_token in a way that
matches the dataset tokenizer, with the following priority:
Command line argument > model default (if one exists) > "<|endoftext|>".

Additionally, the command line arguments --eot '' and --pad ''
are now correctly interpreted as referring to the default
eos_token and pad_token, rather than being interpreted literally as '',
which would in most cases resolve to another special token internally.

The resolution of a correct pad_token (with an ID matching the padding
token used in the tokenized dataset) is necessary for attention masking
to work correctly during training.

### 16-bit model loading
Explicitly casting models to half-precision as model.half() is
not preferred for mixed-precision training, and throws an error
during training if attempted without DeepSpeed.
It doesn't appear to even save VRAM with DeepSpeed enabled,
so this cast is removed.

### Misc.
Additionally, no_init() is changed to a context manager that covers
more steps of model instantiation.