fix(finetuner): Reduce overzealous masking of end-of-sentence tokens
In the default finetuner configuration, both the end-of-sentence
and padding tokens default to <|endoftext|>.
This makes padding identification ambiguous when generating
attention masks.
Ideally, this would be solved by emitting attention masks in gpt_bpe,
but in the absence of that functionality, this commit adds heuristics
such that if the padding and EOS tokens are indistinguishable:
- Only the final context in the dataset is assumed to have padding.
  - It will have EOS masked as padding; all others preserve EOS tokens.
- The prompt file evaluator will assume <|endoftext|> is always EOS.

This should enable the finetuner to learn <|endoftext|> as EOS again,
which may be helpful for training on short-form content.