feat(finetuner): Validate arguments, annotate valid ranges in workflow
This adds validation on every numeric parameter to the finetuner and
evaluator, as well as early validation for some file parameters.
The program will exit early if a constraint is violated with
a descriptive parser error message stating why the parameter is invalid.

For example, `--zero-stage` must be between 0 and 3, `--top-p` and
`--train-ratio` are bounded in the range [0, 1], and `--bs` is either
strictly positive or exactly -1.
These are loose bounds intended to catch simple errors for people
unfamiliar with the meanings of all the parameters.
The ranges are also annotated in comments in the finetune-workflow
for educational purposes.

File validation checks that the `--dataset` file exists, that the
`--ds-config` and `--prompt-file` file paths are valid
if they are provided at all, and that the `--prompt-file` is readable
and its contents aren't blank. These would normally cause errors later
in the finetuner, but checking them early allows the program to emit
more descriptive and helpful errors.

Most of the validation code is organized in a mini-module
called `validation` in `utils.py`, so it doesn't clutter
the global namespace.