fix(finetuner): Pre-compile dependencies without using `-march=native`
DeepSpeed's build process forces in the compile flag `-march=native` and
potentially the preprocessor definition `-D__AVX512__` depending on
the system the Dockerfile is built on. This can lead to errors if
it is built on a system with too new of a CPU, since it might include
instructions not available on nodes' CPUs. This replaces occurrences
of `-march=native` with `-march=skylake` to make the builds more
reproducible, and replaces the preprocessor definition `-D__AVX512__`
with `-D__AVX256__` since the CPUS on most nodes don't support AVX512
instructions yet.