fix(finetuner): Normalize DeepSpeed's compilation to use `-D__AVX256__`
DeepSpeed emits one of three SIMD flags based on the machine it was
compiled on, `-D__AVX512__`, `-D__AVX256__`, or `-D__SCALAR__`, in
decreasing order of processor capability. Currently, AVX512 is switched
to AVX256, but this still leads to different builds if the build host
happens to have an old enough CPU to end up with `-D__SCALAR__`.
This change then further replaces `-D__SCALAR__` with `-D__AVX256__`
to increase the efficiency of the built images in that scenario,
and to make them better match builds performed with newer CPUs.